Support yaml config files & remove HTTP client
