+++ tox.ini
@@ -14,7 +14,7 @@ deps =
     nose
     pytest
 commands =
-    flake8 . --count --select=E901,E999,F821,F822,F823 --show-source --statistics
+    flake8 . --exclude tmp --count --select=E901,E999,F821,F822,F823 --show-source --statistics
     pytest {posargs}
 passenv = WIOTP_API_KEY WIOTP_API_TOKEN WIOTP_ORG_ID
